Set Up for Keras
This commit has the foundational code to transform an individual COA label to a single line of text with a blank background for Keras Training.

Next steps is to create COA Label Strip segmentation, allowing for a automated process of converting each COA label AS WELL AS the foundation for the OCR package itself.